refactor: extract things to files
