+++ raspberrypi_central/webapp/app/standalone/mqtt/mqtt_run.py
@@ -7,7 +7,7 @@ os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'hello_django.settings')
 django.setup()
 
 
-from standalone.mqtt import mqtt_factory  # noqa: E402
+from utils.mqtt import mqtt_factory  # noqa: E402
 from alarm.mqtt import register  # noqa: E402
 
 mqtt = mqtt_factory(client_id='hello-world')
